+++ internal/provider/provider.go
@@ -127,6 +127,7 @@ func (p *MetabaseProvider) Resources(ctx context.Context) []func() resource.Reso
 		NewGlossaryTermResource,
 		NewPermissionsGraphResource,
 		NewPermissionsGroupResource,
+		NewSnippetResource,
 		NewTableResource,
 	}
 }